move ritual util back and sort out util/helper
