+++ src/datagen/java/wayoftime/bloodmagic/datagen/content/datamap/BloodRuneData.java
@@ -9,7 +9,7 @@ import wayoftime.bloodmagic.common.block.BMBlocks;
 import wayoftime.bloodmagic.common.datamap.BMDataMaps;
 import wayoftime.bloodmagic.common.datamap.BloodRune;
 import wayoftime.bloodmagic.api.altar.EnumRuneType;
-import wayoftime.bloodmagic.util.helper.BlockWithItemHolder;
+import wayoftime.bloodmagic.util.blockitem.BlockWithItemHolder;
 
 import java.util.List;
 import java.util.function.Function;
