+++ src/main/java/wayoftime/bloodmagic/common/block/BMBlocks.java
@@ -17,9 +17,9 @@ import wayoftime.bloodmagic.common.datacomponent.BMDataComponents;
 import wayoftime.bloodmagic.common.datamap.BMDataMaps;
 import wayoftime.bloodmagic.common.datamap.BloodRune;
 import wayoftime.bloodmagic.api.altar.EnumRuneType;
-import wayoftime.bloodmagic.util.helper.BlockEntityHelper;
-import wayoftime.bloodmagic.util.helper.BlockWithItemHolder;
-import wayoftime.bloodmagic.util.helper.BlockWithItemRegister;
+import wayoftime.bloodmagic.util.BlockEntityHelper;
+import wayoftime.bloodmagic.util.blockitem.BlockWithItemHolder;
+import wayoftime.bloodmagic.util.blockitem.BlockWithItemRegister;
 
 import java.util.HashMap;
 import java.util.List;
