+++ src/main/java/wayoftime/bloodmagic/common/block/BloodAltarBlock.java
@@ -23,7 +23,7 @@ import net.neoforged.neoforge.items.ItemStackHandler;
 import org.jetbrains.annotations.Nullable;
 import wayoftime.bloodmagic.common.blockentity.BMTiles;
 import wayoftime.bloodmagic.common.blockentity.BloodAltarTile;
-import wayoftime.bloodmagic.util.helper.BlockEntityHelper;
+import wayoftime.bloodmagic.util.BlockEntityHelper;
 
 public class BloodAltarBlock extends Block implements EntityBlock {
     public BloodAltarBlock() {
